Implement GraphQL queries and mutations to manage payments
It implements in Django the payment with QMoney. It uses the module `qmoney`
handling the workflow. This includes:
- a model of the `QmoneyPayment` to persist the info about the ongoing payment
  transaction,
- an unmanaged model of the `Policy` to check the relation with without importing
  the whole `policy` module
- 2 GraphQL queries `qmoneyPayment` and `qmoneyPayments` to respectively
  retrieve and list existing `QmoneyPayment`, and
- 2 GraphQL synchronous mutations that do not use `OpenIMISMutation` from the
  `core` module to keep the coupling low and not have to configure half of
  the apps and dependencies introduced in the `core` module (such as
  `apscheduler`).

This version gives a first foundation covering the full use case (mainly for the
happy path) and raise some important questions such as the way to represent
the amount of money or the possibility to initiate a new payment when one is
still waiting for confirmation for a same given policy.