Fix failing tests verifying payment
When running all the tests, the ones testings the API endpoint `getMoney`
causes a lot of emails to be sent. Due to the amount but also the time needed
for the email to appear in the GMail inbox, it doesn't allow the fixture
`gmail_client` to mark unread all those emails. So, a few emails are still
unread and it makes the test to retrieve not the right email message.

With this change, we make sure that we mark as read all the emails sent by the
tests `TestQmoneyAPIGetMoney` when we not filter out explicitly the tests marked
`with_gmail`.