Integrate the standalone app in the OpenIMIS Django Project
The present Django standalone app has been developed aside the OpenIMIS project.
To keep the coupling low, the tests have been written to be run without the
Django test harness of the OpenIMIS project, and the required models of other
apps (e.g. policy) have been mocked in the present project. Also, we use
`pytest` to run the test "locally".

Those considerations are challenges in themselves:

* how to load dynamically the extern model (from another app) or its mock
  depending on the context?
* how to have tests compatible with two different test frameworks (`pytest` and
  `unittest` + Django) in two different contexts (standalone Django app or the
  whole OpenIMIS Django project and real data in the DB)?

As `pytest` is able to run `unittest` tests, we removed everything specific
to `pytest` in order to make the tests compatible with the Django test harness
(especially make them discoverable). This is mainly about the fixtures.

For the mocked model, we have created a very simple version of the external one
(here `policy.Policy`). Then, we try to load and return the external one. If it
fails we return the local mocked version. This is done with the function
`get_policy_model()`. The generated migration has also been adapted to take in
account the context.

Finally, we have removed the dependencies `pytest-dotenv` and have developed
a simple custom implementation that works in both cases (see `env.py`).