Create Premium and Activate Policy when payment has been proceeded
So far, the implementation was completely decoupled from the policy domain
logic: a proceeded payment didn't change anything, when it should 1. create
a premium to track the payment, and 2. activate the policy.

In the contribution module, there's a function that handles and encapsulates
that logic: `contribution.gql_mutations.update_or_create_premium`. The present
changes will call it once a payment has been proceeded.

As we did for Policy, we use an approach that allows to keep the module
decoupled for the rest, i.e. mocked model dynamically used in the context of
"standalone app" (such as when running tests with pytest without anything else),
or real models dynamically used in the context of the whole OpenIMIS.