Check if user is authenticated and authorized
When querying a list or a given Qmoney payment, requesting a new one, or
proceeding it, it will now check that the user is correctly authenticated and
authorized.

A permission is defined for each GQL query and mutation (get, list, request, and
proceed).

To test, a context is built with the user and used to execute the query. If the
tests are run with the OpenIMIS test harness or standalone/isolated, we use
respectively the test helpers of the OpenIMIS core module or a simple stub.

Aside, the translation has been added. A lot of texts still need to be
translated, but it's a start.