Improve the loading of external models
As the module relies on models coming from other modules (policy, contribution),
we implemented a dynamic discovery of those with a fallback to fake models.
Those fake models allow us to mimic the contract with the real ones, and are
useful during the tests. But the fake models were present as Django models and
could be run in production if someone made an unfortunate change in the present
code.

To avoid such situation, the loading has been improved and relies now on a
Django setting `CUSTOM_MODELS`. When the external models aren't found, it will
check that setting and dynamically load the class with the provided module and
class names. The fake models live now in the tests directory and are not loaded
as Django models by default. Also, the setting is empty by default.

Finally, some tests check that this loading system behaves as expected as
a standalone app or integrated with the whole OpenIMIS app.