Track activity with MutationLog
OpenIMIS tracks submitted mutation with a model named `MutationLog`. This is
done through a layer allowing to asynchronously processed mutations. Due to the
complexity of this layer, we haven't reused that layer. Still, the activity
audit makes sense.

So we create a MutationLog for each GQL mutation (proceed and request a Qmoney
payment). It is marked as a success or an error according to the result of the
mutation.

We use the same dynamic loading of external models than for policy and premium.