Extract all QMoney API classes into a dedicated module
Everything managing the interaction with the QMoney API was in the module
`qmoney_payment.qmoney`. This could led to some confusion due to the naming.

To make things clear, we have extracted everything in a dedicated module
`qmoney_payment.api`. Each class is now in its respective module. And we have
renamed the facade as `Client` from `QMoney`.